fix: drop redundant toLong() conversion in dayKeyOf
toEpochDays() already returns Long; the extra call was flagged by the
compiler and tracked as a final-review cleanup item.

+++ composeApp/src/commonMain/kotlin/fr/dayview/app/Detours.kt
@@ -59,7 +59,7 @@ fun pushRecentDetourMotif(recents: List<String>, motif: String): List<String> {
 fun dayKeyOf(
     now: Instant,
     timeZone: TimeZone = TimeZone.currentSystemDefault(),
-): Long = now.toLocalDateTime(timeZone).date.toEpochDays().toLong()
+): Long = now.toLocalDateTime(timeZone).date.toEpochDays()
 
 /** A distraction source: episodes grouped by normalized motif. */
 data class DetourSource(val label: String, val colorIndex: Int, val total: Duration)
